241024_연상훈_회원가입오류_해결

+++ src/test/java/com/example/actionprice/user/UserRepositoryTest.java
@@ -37,8 +37,10 @@ public class UserRepositoryTest {
     if(user != null){
       System.out.println("the user exists. and we delete the user");
       RefreshTokenEntity refreshToken = user.getRefreshToken();
-      refreshToken.setUser(null);
-      refreshTokenRepository.delete(refreshToken);
+      if(refreshToken != null){
+        refreshToken.setUser(null);
+        refreshTokenRepository.delete(refreshToken);
+      }
       userRepository.delete(user);
       return;
     }
